requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai>=1.35.0
pinecone-client>=3.2.2
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
requests>=2.32.3
